ursys: move tests into their own directory to help with type bundling with dts-bundle-generator
